requirements.txt
python-telegram-bot==13.15
pymongo==4.3.3
qrcode==7.4.2
pillow==9.5.0
python-dotenv==1.0.0
schedule==1.2.0
pandas==2.0.1
matplotlib==3.7.1
pyzbar==0.1.9
urllib3==1.26.15
six==1.16.0
certifi==2023.7.22
